requirements.txt
confluent-kafka
python-dateutil

